Rename components/DevelopmentModeMark.vue to components/_shared/DevelopmentModeMark.vue
