+++ layouts/default.vue
@@ -36,8 +36,8 @@
 import { Component, Vue } from 'nuxt-property-decorator'
 import ScaleLoader from 'vue-spinner/src/ScaleLoader.vue'
 
+import DevelopmentModeMark from '@/components/_shared/DevelopmentModeMark.vue'
 import SideNavigation from '@/components/_shared/SideNavigation.vue'
-import DevelopmentModeMark from '@/components/DevelopmentModeMark.vue'
 import NoScript from '@/components/NoScript.vue'
 import Data from '@/data/data.json'
 import { convertDateToSimpleFormat } from '@/utils/formatDate'
